Finalizing Categories functionalities
I think its all done, check if maybe needs a few things here and there.
I tested adding empty categories or any categories to Products, and
category deletion removing the category from all Products, all working.

TODO
An idea for super later in the future (Not now):
add a field called 'role-lock' to Categories
when a user presses a button to 'role-lock' a category, it will add a
'role-lock' of the userLevel of that user. This will add an extra check
in our backend when a user tries to manage this category, and now only
users with that userLevel of the role lock or above can manage this
category